+++ packages/react-scripts/config/webpack.config.js
@@ -409,7 +409,7 @@ module.exports = function(webpackEnv, options = {}) {
         // Make sure your source files are compiled, as they will not be processed in any way.
         new ModuleScopePlugin(paths.appSrc, [
           paths.appPackageJson,
-          ...glob.sync(path.join(paths.appPath, '/*.md')),
+          ...glob.sync(path.join(paths.appPath, '/*.md'), { realpath: true }),
         ]),
       ],
     },
@@ -472,10 +472,39 @@ module.exports = function(webpackEnv, options = {}) {
             // Process application JS with Babel.
             // The preset includes JSX, Flow, TypeScript, and some ESnext features.
             {
-              test: /\.(js|mjs|jsx|ts|tsx)$/,
+              test: /\.(js|mjs|jsx)$/,
               include: paths.appSrc,
-              loader: require.resolve('babel-loader'),
-              options: getBabelLoaderOptions(),
+              use: [
+                {
+                  loader: require.resolve('babel-loader'),
+                  options: getBabelLoaderOptions(),
+                },
+              ],
+            },
+            {
+              test: /\.(ts|tsx)$/,
+              include: paths.appSrc,
+              use: [
+                {
+                  loader: require.resolve('babel-loader'),
+                  options: getBabelLoaderOptions(),
+                },
+                !isEnvLib && {
+                  loader: require.resolve('react-docgen-typescript-loader'),
+                  options: {
+                    tsconfigPath: paths.appTsConfig,
+                    docgenCollectionName: null,
+                    savePropValueAsString: true,
+                    propFilter: prop => {
+                      if (prop.parent) {
+                        // filter out external types
+                        return !prop.parent.fileName.includes('node_modules');
+                      }
+                      return true;
+                    },
+                  },
+                },
+              ].filter(Boolean),
             },
             // Process any JS outside of the app with Babel.
             // Unlike the application JS, we only compile the standard ES features.