+++ packages/react-scripts/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @lighting-beetle/lighter-react-scripts
 
+## 3.8.2
+
+### Patch Changes
+
+- d87d58f6: Filter out all external prop types (TS-only)
+
+## 3.8.1
+
+### Patch Changes
+
+- 6985ee1e: Fix imports outside `/src` not working in windows
+- 386ef8fe: Add propFilter to filter out unwanted type defs, i.e. all from `@types/react`.
+
+## 3.8.0
+
+### Minor Changes
+
+- 49921a80: add support form generating react-docgen info from TS prop definitions
+
 ## 3.7.0
 
 ### Minor Changes